+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - server/mssql: fix malformed JSON answer on empty tables
 - server/mssql: fix runtime errors when selecting geography/geometry columns
+- server/mssql: supports connection pooling to sql server
 
 ## v1.4.0-alpha.2
 ### MSSQL support